# src/main/java/ru/netology/cloudstoragediplom/dto/file/FileDto.java
package ru.netology.cloudstoragediplom.dto.file;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class FileDto {
    private String hash;
    private byte[] file;

}